+ gameobject: new function SetInWorldTile to replace the old behaviour: set x,y,z -> updatescreenpos(), addtotile() + effects: int to ushort/sbyte position

+++ src/Game/GameObjects/MovingEffect.cs
@@ -48,12 +48,12 @@ namespace ClassicUO.Game.GameObjects
             EffectManager manager,
             uint src,
             uint trg,
-            int xSource,
-            int ySource,
-            int zSource,
-            int xTarget,
-            int yTarget,
-            int zTarget,
+            ushort xSource,
+            ushort ySource,
+            sbyte zSource,
+            ushort xTarget,
+            ushort yTarget,
+            sbyte zTarget,
             ushort graphic,
             ushort hue,
             bool fixedDir,
@@ -174,7 +174,7 @@ namespace ClassicUO.Game.GameObjects
             if (newX != sX || newY != sY)
             {
                 // TODO: Z is wrong. We have to calculate an average
-                SetSource(newX, newY, sZ);
+                SetSource((ushort) newX, (ushort) newY, (sbyte)sZ);
 
                 Vector2 nextSource = new Vector2((newCoordX - newCoordY) * 22, (newCoordX + newCoordY) * 22 - offsetSourceZ * 4);
